refactor(web): move component tests into sibling __tests__ directories
