+++ web/app/components/__tests__/browser-initializer.spec.ts
