+++ web/app/components/base/amplitude/__tests__/AmplitudeProvider.spec.tsx
@@ -2,7 +2,7 @@ import * as amplitude from '@amplitude/analytics-browser'
 import { sessionReplayPlugin } from '@amplitude/plugin-session-replay-browser'
 import { render } from '@testing-library/react'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
-import AmplitudeProvider, { isAmplitudeEnabled } from './AmplitudeProvider'
+import AmplitudeProvider, { isAmplitudeEnabled } from '../AmplitudeProvider'
 
 const mockConfig = vi.hoisted(() => ({
   AMPLITUDE_API_KEY: 'test-api-key',
