+++ web/app/components/base/markdown-with-directive/components/__tests__/markdown-with-directive-schema.spec.ts
@@ -1,4 +1,4 @@
-import { validateDirectiveProps } from './markdown-with-directive-schema'
+import { validateDirectiveProps } from '../markdown-with-directive-schema'
 
 describe('markdown-with-directive-schema', () => {
   beforeEach(() => {
